fix(vehicle_new): correct systematic estimator sign errors
Changes:
- complementary_estimator: Mahony correction cross order fixed to
  meas_down x est_down (was est x meas = positive feedback diverging
  to the inverted equilibrium; verified by derivation and simulation)
- eskf_core: F[vel][att] block sign fixed to -R[a_c x]dt per the
  right-perturbation convention (q = q_hat (x) dq); the flipped sign
  inverted the vel-att cross covariance so flow-velocity updates
  corrected attitude backwards
- eskf_core: setAttitudeFromGravity roll/pitch signs fixed
  (roll = atan2(-ay,-az), pitch = atan2(ax, sqrt(ay^2+az^2)) for the
  specific-force convention f = -gravity_body at rest)
- calibration: computeLevelOffset no longer double-adds G (az is the
  raw rest mean = bias - G) and uses the same specific-force atan2
  convention as setAttitudeFromGravity
- eskf_core: removed fprintf(stderr) chi2 debug instrumentation from
  the 400Hz vectorUpdate3 path (would block on UART on real hardware)
- eskf_core: updateToFVelocity prev-height moved from function-local
  static to members, discarded by reset() (no re-takeoff spike)
- eskf_estimator: flow dt computed from uint32 timestamps with
  wrap-safe unsigned subtraction (float lost us resolution after
  ~17 min; >1s gaps re-seed with nominal dt)

Root cause pattern: specific force f vs gravity g sign confusion in
3 places (Mahony / setAttitudeFromGravity / computeLevelOffset); the
convention is now documented at each site.

Verification:
- host tests 18/18 PASS
- SIL pos_flight.scn PASS on vehicle_new (att_rmse 0.64 deg,
  horizontal_drift_max 0.97 m, all gates green)

Next steps:
- Batch 3: failsafe/state-machine fixes (COMM_LOST airborne-wide,
  impact latch reset on disarm, battery latch split, ESKF_DIVERGED
  actually publishing estimator reset)
- Full SIL regression + sf build vehicle_new after all batches

+++ firmware/vehicle_new/components/sf_estimator_eskf/include/eskf_estimator.hpp
@@ -45,7 +45,14 @@ private:
     StateEstimate convertState(uint32_t timestamp) const;
     EskfCore core_;
     StateEstimate cached_state_ = {};
-    float last_flow_time_ = 0;
+    // Last flow timestamp [us], kept as uint32 (NOT float): a float loses µs
+    // resolution after ~17 min of uptime (24-bit mantissa), corrupting the flow
+    // dt and thus the flow-velocity observation. Unsigned subtraction handles
+    // the ~71.6 min uint32 wrap correctly. 0 = no sample yet.
+    // 最終フロー時刻 [us]。float でなく uint32 で保持: float は稼働約17分で µs 分解能を
+    // 失い（仮数24bit）、フロー dt＝速度観測が劣化する。符号なし減算は約71.6分の uint32
+    // ラップも正しく扱える。0 = 未受信。
+    uint32_t last_flow_time_ = 0;
 };
 
 }  // namespace sf
